+++ test/test_density.py
@@ -7,8 +7,8 @@ def test():
     assert elements.D.density == elements.H.density * elements.D.mass/elements.H.mass
 
     # check number density and unit cell width
-    assert abs(elements.Ca.interatomic_distance - 3.50166387163) < 1e-10
-    assert abs(elements.Fe.number_density - 8.4910635606518029e+22) < 1e12
+    assert abs(elements.Ca.interatomic_distance - 3.5016640712645786) < 1e-10
+    assert abs(elements.Fe.number_density - 8.491062108378546e+22) < 1e12
     assert elements.Ca.interatomic_distance_units == "angstrom"
     assert elements.Fe.number_density_units == "1/cm^3"